6-Hypothetical-Document-Embedding-HyDE/pyproject.toml
[project]
name = "6-hypothetical-document-embedding-hyde"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.1.4",
    "langchain-qdrant>=0.2.0",
    "openai>=1.82.0",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
]
